aggregator: separate out status controller metrics
