+++ staging/src/k8s.io/kube-aggregator/pkg/controllers/status/metrics/metrics_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apiserver
+package metrics
 
 import (
 	"strings"
